requirements.txt
pandas
scikit-learn
joblib
apache-airflow==2.10.2
streamlit
